Add speed scaling and  endgame slide

+++ tObject.h
@@ -1,7 +1,6 @@
 #ifndef INC_2SEMESTR_TOBJECT_H
 #define INC_2SEMESTR_TOBJECT_H
 #include <SFML/Graphics.hpp>
-#include <filesystem>
 #include <cmath>
 #include "tBase_Obj.h"
 
@@ -9,9 +8,9 @@
 class tObject : virtual public tBase_Obj {
 
 public:
-    tObject(int _X, int _Y, int _Width, int _Height);
+    tObject(float _X, float _Y, float _Width, float _Height);
     ~tObject();
-    void Move(int dX, int dY);
+    void Move(float dX, float dY);
 };
